requirements.txt
agno>=1.2.8
streamlit>=1.44.1
firecrawl-py>=0.1.0
elevenlabs>=0.2.0
google-genai>=0.1.0
requests>=2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2